Move model classes CamelKafkaConnectorModel and CamelKafkaConnectorOptionModel in their own module
