+++ tooling/camel-kafka-connector-generator-maven-plugin/pom.xml
@@ -136,11 +136,10 @@
             <version>${camel.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-tooling-model</artifactId>
-            <version>${camel.version}</version>
-        </dependency>
-        
+            <groupId>org.apache.camel.kafkaconnector</groupId>
+            <artifactId>camel-kafka-connector-model</artifactId>
+            <version>${project.version}</version>
+        </dependency>        
 
         <!-- jandex for annotation processing -->
         <dependency>
