+++ tooling/camel-kafka-connector-model/src/main/java/org/apache/camel/kafkaconnector/model/CamelKafkaConnectorOptionModel.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.kafkaconnector.maven.dto;
+package org.apache.camel.kafkaconnector.model;
 
 import org.apache.camel.tooling.model.Strings;
